refactor: api 함수 카테고리 별로 분리 후 적용
