+++ pages/api/dashboards.ts
@@ -1,9 +1,5 @@
 import AXIOS from '@/lib/axios'
-
-interface GetPagenationType<T> {
-  data: T[]
-  totalCount: number
-}
+import { GetPagenationType } from '@/src/types/mydashboard'
 
 export const fetchGetDashboards = async <U>(
   page: number,
@@ -42,24 +38,4 @@ export const fetchGetInviteeEmails = async <U>(
     data: { invitations, totalCount },
   } = response
   return { data: invitations, totalCount }
-}
-
-export const fetchGetDashboardMembers = async <U>(
-  page: number,
-  dashboardId: number,
-  visibleDataNum: number,
-): Promise<GetPagenationType<U>> => {
-  const token = localStorage.getItem('accessToken')
-  const response = await AXIOS.get(
-    `/members?page=${page}&size=${visibleDataNum}&dashboardId=${dashboardId}`,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    },
-  )
-  const {
-    data: { members, totalCount },
-  } = response
-  return { data: members, totalCount }
 }
